pyproject.toml
[project]
name = "GenericAIApp"
version = "0.1.0"
description = "Generic AI App that contains chat history management, mutli agent orchestration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Agent Framework (Microsoft)
    "agent-framework>=1.0.0b251120",
    # Azure
    "azure-identity>=1.15.0",
    "azure-keyvault-secrets>=4.8.0",
    "azure-storage-blob>=12.19.0",
    # FastAPI (async API)
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic-settings>=2.0.0",
    # Flask UI (legacy, can be removed when FastAPI migration complete)
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "gunicorn>=21.2.0",
    "gevent>=24.2.1",
    # Storage backends
    "psycopg2-binary==2.9.9",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    # Config
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "azure-monitor-opentelemetry-exporter>=1.0.0b45",
]


[tool.uv]
prerelease = "allow"
